# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezycrowd AI crowd detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Detection {
  x: number;
  y: number;
  width: number;
  height: number;
  confidence: number;
}

export interface DetectionResult {
  sessionId: string;
  /** Number of people detected */
  peopleCount: number;
  /** Whether people count exceeded threshold */
  alertTriggered: boolean;
  /** Alert threshold used for this detection */
  threshold: number;
  /** Average detection confidence (0.0-1.0) */
  confidence: number;
  /** Time taken to process the file in milliseconds */
  processingTimeMs: number;
  timestamp: string;
  /** URL to annotated result image (if available) */
  imageUrl?: string;
  detections?: Detection[];
}

export interface ThresholdConfig {
  /**
   * Number of people that triggers an alert
   * @minimum 1
   */
  threshold: number;
}

export interface Session {
  id: string;
  peopleCount: number;
  alertTriggered: boolean;
  threshold: number;
  confidence: number;
  processingTimeMs: number;
  createdAt: string;
  fileName?: string;
}

export interface SessionList {
  sessions: Session[];
  total: number;
}

export interface Alert {
  id: string;
  sessionId: string;
  peopleCount: number;
  threshold: number;
  createdAt: string;
}

export interface AlertList {
  alerts: Alert[];
  total: number;
}

export interface StatsSummary {
  totalSessions: number;
  totalAlerts: number;
  averagePeopleCount: number;
  maxPeopleCount: number;
  /** Percentage of sessions that triggered alerts (0.0-1.0) */
  alertRate: number;
  recentSessions?: Session[];
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type DetectCrowdBody = {
  /** Image or video file to analyze */
  file: Blob;
  /** People count threshold for alert (default 10) */
  threshold?: number;
};

export type ListSessionsParams = {
  limit?: number;
  alert_only?: boolean;
};

export type ListAlertsParams = {
  limit?: number;
};