# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ezycrowd AI crowd detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /detect:
    post:
      operationId: detectCrowd
      tags: [detect]
      summary: Detect crowd in uploaded image or video frame
      description: Upload an image or video file; returns people count, bounding boxes, and alert status
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image or video file to analyze
                threshold:
                  type: integer
                  description: People count threshold for alert (default 10)
                  default: 10
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionResult"
        "400":
          description: Invalid file type or missing file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Detection processing error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /detect/threshold:
    get:
      operationId: getThreshold
      tags: [detect]
      summary: Get current alert threshold
      responses:
        "200":
          description: Current threshold value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThresholdConfig"
    put:
      operationId: updateThreshold
      tags: [detect]
      summary: Update alert threshold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThresholdConfig"
      responses:
        "200":
          description: Updated threshold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThresholdConfig"
...
components:
  schemas:
...
    DetectionResult:
      type: object
      required:
        - sessionId
        - peopleCount
        - alertTriggered
        - threshold
        - confidence
        - processingTimeMs
        - timestamp
      properties:
        sessionId:
          type: string
        peopleCount:
          type: integer
          description: Number of people detected
        alertTriggered:
          type: boolean
          description: Whether people count exceeded threshold
        threshold:
          type: integer
          description: Alert threshold used for this detection
        confidence:
          type: number
          format: float
          description: Average detection confidence (0.0-1.0)
        processingTimeMs:
          type: number
          format: float
          description: Time taken to process the file in milliseconds
        timestamp:
          type: string
          format: date-time
        imageUrl:
          type: string
          description: URL to annotated result image (if available)
        detections:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
    Detection:
      type: object
      required: [x, y, width, height, confidence]
      properties:
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        confidence:
          type: number
    ThresholdConfig:
      type: object
      required: [threshold]
      properties:
        threshold:
          type: integer
          minimum: 1
          description: Number of people that triggers an alert
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string