# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezycrowd AI crowd detection API
 * OpenAPI spec version: 0.1.0
 */

export interface Detection {
  x: number;
  y: number;
  width: number;
  height: number;
  confidence: number;
}