# lib/api-zod/src/generated/types/detectionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezycrowd AI crowd detection API
 * OpenAPI spec version: 0.1.0
 */
import type { Detection } from "./detection";

export interface DetectionResult {
  sessionId: string;
  /** Number of people detected */
  peopleCount: number;
  /** Whether people count exceeded threshold */
  alertTriggered: boolean;
  /** Alert threshold used for this detection */
  threshold: number;
  /** Average detection confidence (0.0-1.0) */
  confidence: number;
  /** Time taken to process the file in milliseconds */
  processingTimeMs: number;
  timestamp: Date;
  /** URL to annotated result image (if available) */
  imageUrl?: string;
  detections?: Detection[];
}