# lib/api-zod/src/generated/types/thresholdConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezycrowd AI crowd detection API
 * OpenAPI spec version: 0.1.0
 */

export interface ThresholdConfig {
  /**
   * Number of people that triggers an alert
   * @minimum 1
   */
  threshold: number;
}